Make detect_diff_type accept a single line
Instead of iterating through the inputstream array each time when searching for
a diff type, check only a newly added line. Put the iteration responsibility
into the caller for the diffy case. Note that the body of the foreach-loop has
not changed, it is only indented to the left after removing the foreach.

While we are at it, prevent the global diff_type from being modified by
detect_diff_type (just return it, the caller will set it anyway).