Start processing data before reading input with diffy
This was an attempt to recognize `svn log --diff` better, but that failed here.
Anyway, the changes made here avoid looping through the input data twice at the
cost of some additional complexity in the inner loop. It also allows data to be
processed as soon as it is available instead of reading everything in memory
first before processing data.