Detect earlier that the format is not sdiff
Count the columns that are possibly a marker for sdiff. If any column (ignoring
CVS and diff command lines) does not contain a marker, the column is not
considered a marker anymore.

(The following is skipped when --difftype is set)
Whenever no marker columns are found, if another diff type can be recognized,
that one is used.

This fixes blocking behavior of `svn log --diff | colordiff`.